[change] move map_nearest to KNN helper file
